+++ lib/src/controls/form/auto_suggest_box.dart
@@ -244,12 +244,15 @@ class AutoSuggestBox<T> extends StatefulWidget {
   final TextStyle? style;
 
   /// Controls the [BoxDecoration] of the box behind the text input.
-  final BoxDecoration? decoration;
+  ///
+  /// Defaults to having a rounded rectangle grey border and can be null to have
+  /// no box decoration.
+  final WidgetStateProperty<BoxDecoration>? decoration;
 
   /// Controls the [BoxDecoration] of the box in front of the text input.
   ///
   /// If [highlightColor] is provided, this must not be provided
-  final BoxDecoration? foregroundDecoration;
+  final WidgetStateProperty<BoxDecoration>? foregroundDecoration;
 
   /// The highlight color of the text box.
   ///