+++ lib/src/controls/form/number_box.dart
@@ -186,6 +186,17 @@ class NumberBox<T extends num> extends StatefulWidget {
   /// {@macro flutter.widgets.editableText.showCursor}
   final bool? showCursor;
 
+  /// Controls the [BoxDecoration] of the box behind the text input.
+  ///
+  /// Defaults to having a rounded rectangle grey border and can be null to have
+  /// no box decoration.
+  final WidgetStateProperty<BoxDecoration>? decoration;
+
+  /// Controls the [BoxDecoration] of the box in front of the text input.
+  ///
+  /// If [highlightColor] is provided, this must not be provided
+  final WidgetStateProperty<BoxDecoration>? foregroundDecoration;
+
   /// The highlight color of the text box.
   ///
   /// If [foregroundDecoration] is provided, this must not be provided.
@@ -290,6 +301,8 @@ class NumberBox<T extends num> extends StatefulWidget {
     this.showCursor,
     this.highlightColor,
     this.unfocusedColor,
+    this.decoration,
+    this.foregroundDecoration,
     this.style,
     this.textAlign,
     this.keyboardType = TextInputType.number,
@@ -542,6 +555,9 @@ class NumberBoxState<T extends num> extends State<NumberBox<T>> {
       cursorRadius: widget.cursorRadius,
       cursorWidth: widget.cursorWidth,
       highlightColor: widget.highlightColor,
+      unfocusedColor: widget.unfocusedColor,
+      decoration: widget.decoration,
+      foregroundDecoration: widget.foregroundDecoration,
       prefix: widget.leadingIcon,
       focusNode: focusNode,
       controller: controller,
@@ -549,7 +565,6 @@ class NumberBoxState<T extends num> extends State<NumberBox<T>> {
       enabled: widget.onChanged != null,
       suffix:
           textFieldSuffix.isNotEmpty ? Row(children: textFieldSuffix) : null,
-      unfocusedColor: widget.unfocusedColor,
       style: widget.style,
       textAlign: widget.textAlign ?? TextAlign.start,
       keyboardAppearance: widget.keyboardAppearance,